+++ etils/epath/testing.py
@@ -66,10 +66,14 @@ class _MockBackend(_backend_cls):
   def glob(self, path: PathLike) -> list[str]:
     return self._get_fn('glob')(path)
 
-  def makedirs(self, path: PathLike) -> None:
+  def makedirs(self, path: PathLike, *, exist_ok: bool = False) -> None:
+    if exist_ok and self.exists(path):
+      raise FileExistsError(f'{path} already exists.')
     return self._get_fn('makedirs')(path)
 
-  def mkdir(self, path: PathLike) -> None:
+  def mkdir(self, path: PathLike, *, exist_ok: bool = False) -> None:
+    if exist_ok and self.exists(path):
+      raise FileExistsError(f'{path} already exists.')
     return self._get_fn('mkdir')(path)
 
   def rmtree(self, path: PathLike) -> None: